Refactoring modules structure.
