+++ run.py
@@ -1,19 +1,20 @@
 import logging.config
 
-import SETTINGS
-from src.nuforc.modules.scraping import NUFORCScraper
+from src.nuforc.SETTINGS import DEFAULT_ENGINE_SETTINGS
+from src.nuforc.SETTINGS import LOGGING_CONFIG
+from src.nuforc.scraping import NUFORCScraper
 
-logging.config.dictConfig(SETTINGS.LOGGING_CONFIG)
+logging.config.dictConfig(LOGGING_CONFIG)
 logger = logging.getLogger("root")
 
 
 def execute_scraping():
     scraper = NUFORCScraper(
-        scraping_mode=SETTINGS.SCRAPING_MODE,
-        timespan_start=SETTINGS.TIMESPAN_START,
-        timespan_end=SETTINGS.TIMESPAN_END,
-        n_scraping_retries=SETTINGS.N_SCRAPING_RETRIES,
-        output_folder=SETTINGS.OUTPUT_FOLDER,
+        scraping_mode=DEFAULT_ENGINE_SETTINGS.scraping_mode,
+        timespan_start=DEFAULT_ENGINE_SETTINGS.timespan_start,
+        timespan_end=DEFAULT_ENGINE_SETTINGS.timespan_end,
+        n_scraping_retries=DEFAULT_ENGINE_SETTINGS.n_scraping_retries,
+        output_folder=DEFAULT_ENGINE_SETTINGS.output_folder,
     )
     scraper.scrape()
     scraper.save_events()
