--- src/nuforc/__main__.py
@@ -1,4 +1,0 @@
-import nuforc
-
-if __name__ == '__main__':
-    pass
